Add files via upload
There are too many variables to directly do a pairs analysis or to assess collinearity in a meaningful way. Accordingly, I thought it would be useful to initially run three parallel models, each based on the category of variable involved (disease, monetary ("money"), and weight). Using these, I was able to eliminate some of the predictors. I then started to check assumptions (Shapiro / BP / QQ / Resid vs Fitted) and these show that there are some issues with the data.