Add files via upload
OK, I've added a new "submodel" or "thematic model" and carried its analysis through the same steps as the others. We now have a BIC and an AIC model. It would be good to compare them (a simple ANOVA would do the trick) but I need to call it a night. It's late my time.