Add files via upload
I've improved the subheadings so you can quickly jump to individual sections of the analysis. I've made copious notes on where you can contribute. I have also added "TO DO" markers to indicate what still needs to be done. The TO DOs are not an exhaustive, as there are definitely more things we can do. I'm traveling tomorrow, but hope to get to my destination in time to be able to talk to the professor about issues I see in our data.