numerous edits / additions
Hi Yoga / Zoheb,
So I went through and edited as much as I could stylistically. I also made a couple of minor corrections. Finally I added a brief section acknowledging the discovery of "complex linear relationships". However, there is a weird problem with one of the plots in that section. It looks fine if I simply run the script in the RMarkdown, but the middle (of three) plots does not produce itself correctly in the knit file. Could one of you figure out what the problem is? It's 1:30 am, and I have a long drive early tomorrow morning. Finally, I fleshed out the final discussion. 

Overall I think it looks really great. Thanks for all the additions. I think if you can fix the bug, we should be good to go. (One time it displayed correctly, but the other times it didn't. Must be something simple.

Since I'm traveling and may not have access to the Internet, I recommend one of you makes the final upload. Please confirm via What's App. Thanks.

Daniel