final submission upload
We are identified as "crime data team" so I kept it in the title of the zip file for ease of identification. The professor knows we are working on life expectancy. Feel free to get in touch if you see something wrong. I inadvertently uploaded an HTML file yesterday, so Yoga didn't know I had made so many edits and additions. That's why I am resubmitting.